requirements.txt
PyYaml >= 3.11
suds >= 0.4
zuora-client >= 1.0
